Remove commcare case polling, and the just-in-time update in CommCareCaseLoader
Case forwarder sends us the latest case updates, so we shouldn't need to
update based on the CommCare response.

Transferred cases are an exception: although we do have the latest state
from CommCare, we look up transfers via API calls. We do need to update
the original case with the transferred data.

In the future, we could look up transfers using database information
instead, and avoid any API calls in the CommCareCaseLoader.

0dcf7cca0d7994ca690d9b6961b17ef9c8d5a7b8